+++ readme/RELEASE_NOTES.md
@@ -1,5 +1,7 @@
 * [#42](https://github.com/skuzzle/snapshot-tests/issues/42): Publish flattened version of our artifact to fix BOM distribution
 * [#43](https://github.com/skuzzle/snapshot-tests/issues/43): Allow XML structure comparison for String input
+* Add `XmlSnapshot.withEnableXPathDebugging(...)` and `HtmlSnapshot.withEnableXPathDebugging(...)` to print results of applying custom comparison rules
+* Fixed bugs and performance issues in XPath comparison engine
 
 Maven Central coordinates for this release: